+++ retdec/tools/decompiler.py
@@ -375,6 +375,27 @@ def parse_args(argv):
         help='Decompile all functions, even if they are not reachable.'
     )
     parser.add_argument(
+        '--only-funcs',
+        dest='sel_decomp_funcs',
+        metavar='FUNCS',
+        help='Decompile only the given functions (a comma-separated '
+             'list of function names, e.g. func1,func2).'
+    )
+    parser.add_argument(
+        '--only-ranges',
+        dest='sel_decomp_ranges',
+        metavar='RANGES',
+        help='Decompile only the given address ranges (a comma-separated '
+             'list of address ranges, e.g. 0x100-0x200,0x500-0x600).'
+    )
+    parser.add_argument(
+        '--decoding',
+        dest='sel_decomp_decoding',
+        metavar='TYPE',
+        choices=['everything', 'only'],
+        help='What should be decoded in a seletive decompilation? Default: everything'
+    )
+    parser.add_argument(
         '--no-addresses',
         dest='decomp_emit_addresses',
         action='store_false',
@@ -472,6 +493,9 @@ def main(argv=None):
     add_decompilation_param_when_given(args, params, 'decomp_var_names')
     add_decompilation_param_when_given(args, params, 'decomp_optimizations')
     add_decompilation_param_when_given(args, params, 'decomp_unreach_funcs')
+    add_decompilation_param_when_given(args, params, 'sel_decomp_funcs')
+    add_decompilation_param_when_given(args, params, 'sel_decomp_ranges')
+    add_decompilation_param_when_given(args, params, 'sel_decomp_decoding')
     add_decompilation_param_when_given(args, params, 'decomp_emit_addresses')
     add_decompilation_param_when_given(args, params, 'generate_archive')
     decompilation = decompiler.start_decompilation(**params)